Add files via upload
This is the challenge website: https://www.frontendmentor.io/challenges/interactive-pricing-component-t0m8PIyY8
This challenge was harder than I expected it to be.
The main problems were the responsiveness (some components change their location on the screen according to screen width), and the range progress color.
Making the range progress bar change color in all browsers was hellish and I felt like giving up halfway-through. In Firefox, it was easy: there was a built-in -moz-range-progress pseudo-element that I styled using css. Easy peasy. It was definitely NOT the case with Chrome and Opera! I had to use JavaScript in the end to make it work.
This web app was challenging to code, and I can't say that it didn't frustrate me at some point, but I enjoyed the overall experience.